quick format to modules
Do we want to have a hardy.py to function as the main wrapping functions? 

Also: As for cross-importing:
I imagine that on the list above, each file will Import the files above it, yes? That way we don't have cross-talk? Exception: yNot maybe imports none of them and can be used by all. 

If that's the case, though, we might need someone on the "Outside" of all the functions to be the main Wrapper.
In which case,